error on make

+++ ft_putpointer.c
@@ -12,7 +12,7 @@
 
 #include "ft_printf.h"
 
-int	ft_hexa_base_pointer(uintptr_t ptr)
+int	ft_hexa_base_pointer(unsigned long long int ptr)
 {
 	char	*base16;
 	int		a;
@@ -23,7 +23,7 @@ int	ft_hexa_base_pointer(uintptr_t ptr)
 	return (a);
 }
 
-int	ft_print_ptr(uintptr_t ptr)
+int	ft_print_ptr(unsigned long long int ptr)
 {
 	int	a;
